config: add dependency to pc-ble-driver into config, default off
Add dependency for the pc-ble-driver into comments, so one can
"easily" enable it on the build if needed. Note: the driver itself
works equally on a USB connected dongle or hardwired module in a
board, so it may not make sense to tie it to a specific board at
all. Let's just make it possible for user to choose if the module
is available and needed.